refactor: move fit_model into captest.filters (capdata re-imports)
fit_model and its statsmodels import move to filters.py so FilterRegression
can use it without an import cycle. capdata.py re-imports it (still needed by
predict_capacities' grps.apply(fit_model)), so capdata.fit_model / pvc.fit_model
keep resolving. Drop the now-unused statsmodels import from capdata.py.
